Add misc/ folder
This is a common pattern: the `misc/` folder is available to anyone when they clone the repo, so one can store any personal files related to the project (files which shouldn't be versioned). The `.gitkeep` file is necessary for git to version the folder (as git doesn't keep track of empty folders), however the added lines in `.gitignore` tells git to ignore any other file in the `misc/` folder. So this folder will always be empty when the repo is cloned by someone (except from the `.gitkeep` file).